refactor: split spatial index into world-space SpatialIndex and render-side VisualIndex
