--- src/game/physics/gridSpace/GridSpatialIndex.ts
@@ -1,735 +1,0 @@
-import { type UnionOfAllItemInPlayTypes } from "../../../model/ItemInPlay";
-import { addXyz, type Xyz } from "../../../utils/vectors/vectors";
-import {
-  projectAabbAxes,
-  type ProjectionOnAxes,
-} from "../../render/sortZ/projectAabbCorners";
-import { blockSizePx } from "../mechanicsConstants";
-
-// Enable debug logging and verification for spatial index operations
-const DEBUG_SPATIAL_INDEX = false;
-// Only log operations for these specific item IDs
-const DEBUG_ITEM_IDS = ["head", "heels"];
-
-// these multiplications empirically give good performance in larger rooms.
-// if the cells are too fine, the search for shadow-casters has to visit
-// many cells - best to keep this small. Nothing else seems to mind
-// the multiplication very much perf-wise, since the number of items in any
-// given cell remains very small
-const cellDepth = blockSizePx.x * 2;
-const cellWidth = blockSizePx.x * 2;
-
-// like above, the multipliers are a balance between each item being in
-// more cells (because smaller cells) and having to do more of the fine
-// comparisons (because cells are big and have a lot of items in them)
-const projCellDepth = blockSizePx.x * 2;
-const projCellWidth = blockSizePx.x * 2;
-
-// two cell coords packed into one 32-bit int (16 bits each) - see #makeCellKey
-type XyCellKey = number;
-
-export type Indexable = {
-  id: string;
-  state: {
-    position: Xyz;
-  };
-  aabb: Xyz;
-  renderAabb?: Xyz;
-  renderAabbOffset?: Xyz;
-};
-
-/**
- * Spatial indexing system that divides the game world into a 2D grid of cells.
- * Each cell contains a Set of items that occupy that space (infinite vertical columns).
- * Uses a sparse Map structure to efficiently handle any coordinate range.
- *
- * Indexing in x,y only is based on the assumption that there will be lots of items
- * overlapping in x and y, but few vertically above or below. The trade off
- * is a simpler index, for considering more cells for collisions. This also allows us to
- * make 'infinitely high' (or just very high) walls since they still don't need to occupy
- * a huge number of 3d cells extending vertically
- */
-export class GridSpatialIndex<
-  RoomId extends string = string,
-  RoomItemId extends string = string,
-  Item extends Indexable = UnionOfAllItemInPlayTypes<RoomId, RoomItemId>,
-  Bin extends Set<Item> = Set<Item>,
-> {
-  /**
-   * Map from cell coordinates (as "x,y" string) to Sets of items in that space
-   * Sparse storage - only cells with items exist
-   */
-  #spatialCells = new Map<XyCellKey, Bin>();
-
-  /**
-   * Map from cell coordinates (as "x,y" string) to Sets of items projected to that
-   * hexagon on screen.
-   * Sparse storage - only cells with items exist
-   */
-  #projectedCells = new Map<XyCellKey, Bin>();
-
-  /**
-   * Map from each item to the set of cell keys it currently occupies.
-   * This reverse mapping is essential for O(1) operations:
-   * - When updating: know which old cells to remove the item from
-   * - When removing: know exactly which cells contain the item
-   * - For existence checks: quickly verify if an item is in the index
-   * Without this, we'd have to search the entire grid for these operations.
-   */
-  #itemToCellKeys = new Map<Item, Set<XyCellKey>>();
-
-  /**
-   * Map from each item to the set of projected cell keys it currently occupies.
-   * This reverse mapping is essential for O(1) operations in projection space:
-   * - When updating: know which old projected cells to remove the item from
-   * - When removing: know exactly which projected cells contain the item
-   * Without this, we'd have to search the entire projection grid for these operations.
-   */
-  #itemToProjectedCellKeys = new Map<Item, Set<XyCellKey>>();
-
-  /**
-   * cache of where items project to on on-screen
-   * axes (x,y,z)(min,max)
-   */
-  #itemToProjectionAxesMap = new Map<Item, ProjectionOnAxes>();
-
-  /**
-   * cached occupied-cell extent of the CUBOID index (x/y cells only - the
-   * index has no z partition), with a dirty flag set on every cuboid mutation.
-   * Recomputed lazily on read by one pass over the cell keys: cheaper than
-   * maintaining min/max incrementally, which would need a full scan on
-   * removal-of-an-extreme anyway. Used to bound the light-beam cell-march so it
-   * cannot walk forever in a room with an open (wall-less) edge.
-   */
-  #cuboidCellExtent:
-    | { minCellX: number; maxCellX: number; minCellY: number; maxCellY: number }
-    | undefined = undefined;
-  #cuboidCellExtentDirty = true;
-
-  constructor(
-    /** Optional iterable of items to initially populate the grid */
-    items?: Iterable<Item>,
-  ) {
-    if (items) {
-      for (const item of items) {
-        this.addItem(item);
-      }
-    }
-  }
-
-  /**
-   * Generic helper to add an item to an index.
-   */
-  #addToIndex(
-    item: Item,
-    cellIterator: Generator<XyCellKey>,
-    itemToCellsMap: Map<Item, Set<XyCellKey>>,
-    ensureCell: (cellKey: XyCellKey) => Bin,
-  ): void {
-    const occupiedCells = new Set<XyCellKey>();
-
-    for (const cellKey of cellIterator) {
-      const cell = ensureCell(cellKey);
-      cell.add(item);
-      occupiedCells.add(cellKey);
-    }
-
-    itemToCellsMap.set(item, occupiedCells);
-  }
-
-  /**
-   * Generic helper to remove an item from an index.
-   */
-  #removeFromIndex(
-    item: Item,
-    cellMap: Map<XyCellKey, Bin>,
-    itemToCellsMap: Map<Item, Set<XyCellKey>>,
-  ): void {
-    const occupiedCells = itemToCellsMap.get(item);
-    itemToCellsMap.delete(item);
-
-    if (DEBUG_SPATIAL_INDEX && DEBUG_ITEM_IDS.includes(item.id)) {
-      console.log(
-        `[GridSpatialIndex] removing item ${item.id} from occupied cells`,
-        occupiedCells,
-      );
-    }
-
-    if (!occupiedCells) {
-      return;
-    }
-
-    for (const cellKey of occupiedCells) {
-      const cell = cellMap.get(cellKey);
-      if (cell) {
-        cell.delete(item);
-        if (cell.size === 0) {
-          cellMap.delete(cellKey);
-        }
-      }
-    }
-  }
-
-  /**
-   * Generic helper to update an item's position in an index.
-   */
-  #updateInIndex(
-    item: Item,
-    cellIterator: Generator<XyCellKey>,
-    cellMap: Map<XyCellKey, Bin>,
-    itemToCellsMap: Map<Item, Set<XyCellKey>>,
-    ensureCell: (cellKey: XyCellKey) => Bin,
-  ): void {
-    const oldCellKeys = itemToCellsMap.get(item);
-    if (!oldCellKeys) {
-      throw new Error(`Item ${item.id} not in index`);
-    }
-
-    const newCellKeys = new Set<XyCellKey>(cellIterator);
-    let deleted = false;
-    let added = false;
-
-    // Remove from old cells that are not in new cells
-    for (const cellKey of oldCellKeys) {
-      if (!newCellKeys.has(cellKey)) {
-        const cell = cellMap.get(cellKey);
-        if (cell) {
-          deleted = true;
-          cell.delete(item);
-        }
-      }
-    }
-
-    // Add to new cells that were not in old cells
-    for (const cellKey of newCellKeys) {
-      if (!oldCellKeys.has(cellKey)) {
-        const cell = ensureCell(cellKey);
-        cell.add(item);
-        added = true;
-      }
-    }
-
-    // Clean up empty cells
-    if (deleted) {
-      for (const cellKey of oldCellKeys) {
-        const cell = cellMap.get(cellKey);
-        if (cell && cell.size === 0) {
-          cellMap.delete(cellKey);
-        }
-      }
-    }
-
-    // Update tracking
-    if (deleted || added) {
-      itemToCellsMap.set(item, newCellKeys);
-    } else {
-      newCellKeys.clear();
-    }
-  }
-
-  /**
-   * Add an item to the grid at its current position.
-   */
-  addItem(
-    /** The item to add to the grid */
-    item: Item,
-  ): void {
-    // Check if item is already in the grid
-    if (this.#itemToCellKeys.has(item)) {
-      const cellKeysEntry = this.#itemToCellKeys.get(item)!;
-      throw new Error(
-        `Item ${item.id} is already in the spatial index at ${[...cellKeysEntry].join(" ")}`,
-      );
-    }
-
-    // Add to cuboid index
-    this.#addToIndex(
-      item,
-      this.#iterateCuboidCellKeys(item.state.position, item.aabb),
-      this.#itemToCellKeys,
-      (cellKey) => this.#ensureCell(cellKey),
-    );
-    this.#cuboidCellExtentDirty = true;
-
-    // Add to projected cell index
-    this.#upsertAxesProjections(item);
-    this.#addToIndex(
-      item,
-      this.#iterateProjectionCellKeys(item),
-      this.#itemToProjectedCellKeys,
-      (cellKey) => this.#ensureProjectedCell(cellKey),
-    );
-
-    if (DEBUG_SPATIAL_INDEX && DEBUG_ITEM_IDS.includes(item.id)) {
-      const spatialCells = Array.from(this.#itemToCellKeys.get(item) || []);
-      const projectedCells = Array.from(
-        this.#itemToProjectedCellKeys.get(item) || [],
-      );
-      console.log(
-        `[GridSpatialIndex] Added item ${item.id}:\n` +
-          `  Position: (${item.state.position.x}, ${item.state.position.y}, ${item.state.position.z})\n` +
-          `  AABB: (${item.aabb.x}, ${item.aabb.y}, ${item.aabb.z})\n` +
-          `  RenderAABB: ${item.renderAabb ? `(${item.renderAabb.x}, ${item.renderAabb.y}, ${item.renderAabb.z})` : "undefined"}\n` +
-          `  RenderAABBOffset: ${item.renderAabbOffset ? `(${item.renderAabbOffset.x}, ${item.renderAabbOffset.y}, ${item.renderAabbOffset.z})` : "undefined"}\n` +
-          `  Spatial cells: [${spatialCells.join(", ")}]\n` +
-          `  Projected cells: [${projectedCells.join(", ")}]`,
-      );
-    }
-  }
-
-  /**
-   * Pack cell coordinates into a single integer key
-   */
-  #makeCellKey(x: number, y: number): XyCellKey {
-    // 16 bits per axis, masked so a negative coord doesn't bleed into the other
-    // half - rooms are far smaller than the ±32767-cell range this allows:
-    return ((x & 0xff_ff) << 16) | (y & 0xff_ff);
-  }
-
-  /** the (x, y) cell coordinates packed into a key by #makeCellKey */
-  #decodeCellKey(key: XyCellKey): { x: number; y: number } {
-    // sign-extend each 16-bit half back to a signed coordinate:
-    return { x: key >> 16, y: (key << 16) >> 16 };
-  }
-
-  /**
-   * Ensure a cell exists, creating it if necessary
-   */
-  #ensureCell(cellKey: XyCellKey): Bin {
-    let cell = this.#spatialCells.get(cellKey);
-    if (!cell) {
-      cell = new Set() as Bin;
-      this.#spatialCells.set(cellKey, cell);
-    }
-    return cell;
-  }
-
-  /**
-   * Ensure a projected cell exists, creating it if necessary
-   */
-  #ensureProjectedCell(cellKey: XyCellKey): Bin {
-    let cell = this.#projectedCells.get(cellKey);
-    if (!cell) {
-      cell = new Set() as Bin;
-      this.#projectedCells.set(cellKey, cell);
-    }
-    return cell;
-  }
-
-  *#iterateCuboidCellKeys(
-    /** The position of the cuboid */
-    position: Xyz,
-    /** The size of the cuboid */
-    size: Xyz,
-  ): Generator<XyCellKey> {
-    const minCorner = position;
-    const maxCorner = addXyz(position, size);
-
-    // Calculate which cells this cuboid occupies (only x,y)
-    const minCellX = Math.floor(minCorner.x / cellWidth);
-    const minCellY = Math.floor(minCorner.y / cellDepth);
-    const maxCellX = Math.floor(maxCorner.x / cellWidth);
-    const maxCellY = Math.floor(maxCorner.y / cellDepth);
-
-    // Yield all cell keys the item occupies (2D only)
-    for (let x = minCellX; x <= maxCellX; x++) {
-      for (let y = minCellY; y <= maxCellY; y++) {
-        yield this.#makeCellKey(x, y);
-      }
-    }
-  }
-
-  #upsertAxesProjections(i: Item): ProjectionOnAxes {
-    const pos =
-      i.renderAabbOffset ?
-        addXyz(i.state.position, i.renderAabbOffset)
-      : i.state.position;
-    const bb = i.renderAabb || i.aabb;
-
-    const existing = this.#itemToProjectionAxesMap.get(i);
-
-    if (existing !== undefined) {
-      // write into the existing object to avoid malloc
-      projectAabbAxes(existing, pos, bb);
-      return existing;
-    }
-    // don't already have in the cache so will have to create a new object
-    const allAxesProjections = projectAabbAxes({}, pos, bb);
-    this.#itemToProjectionAxesMap.set(i, allAxesProjections);
-    return allAxesProjections;
-  }
-
-  /**
-   * project an item to a hexagon, but treat that hexagon as an object 3-dimensional
-   * space, then use 2 of those dimensions to index by (two sets of opposites sides of the hexagon)
-   * Where the projections of the aabb on the three world axes, onto the
-   * screen axes, are taken as the ordinals.
-   *
-   * The hexagons we render to, with a fixed camera angle, are always using the
-   * same slope/angle of their lines (z is vertical, x and y are 2:1 diagonal lines).
-   * Therefore, any two hexagons overlap if they overlap on all three of these axes.
-   * So, it is like a 3d overlapping problem, but for the 2d projections of the 3d axes.
-   */
-  *#iterateProjectionCellKeys(i: Item): Generator<XyCellKey> {
-    const projectionAxes = this.#itemToProjectionAxesMap.get(i);
-
-    if (projectionAxes === undefined) {
-      throw new Error("projectionAxes not calculated for " + i.id);
-    }
-
-    const {
-      xAxisProjectionMin,
-      xAxisProjectionMax,
-      yAxisProjectionMin,
-      yAxisProjectionMax,
-    } = projectionAxes;
-
-    // Calculate which cells this projection occupies (only x,y)
-    const minCellX = Math.floor(xAxisProjectionMin / projCellWidth);
-    const minCellY = Math.floor(yAxisProjectionMin / projCellDepth);
-    const maxCellX = Math.floor(xAxisProjectionMax / projCellWidth);
-    const maxCellY = Math.floor(yAxisProjectionMax / projCellDepth);
-
-    // Yield all cell keys the projection occupies (2D only)
-    for (let x = minCellX; x <= maxCellX; x++) {
-      for (let y = minCellY; y <= maxCellY; y++) {
-        yield this.#makeCellKey(x, y);
-      }
-    }
-  }
-
-  /**
-   * Remove an item from the grid.
-   * Cleans up empty cells to maintain sparse storage.
-   */
-  removeItem(
-    /** The item to remove from the grid */
-    item: Item,
-  ): void {
-    // Check if item is in the grid
-    if (!this.#itemToCellKeys.has(item)) {
-      throw new Error(`Item ${item.id} is not in the spatial index`);
-    }
-
-    if (DEBUG_SPATIAL_INDEX && DEBUG_ITEM_IDS.includes(item.id)) {
-      const spatialCells = Array.from(this.#itemToCellKeys.get(item) || []);
-      const projectedCells = Array.from(
-        this.#itemToProjectedCellKeys.get(item) || [],
-      );
-      console.log(
-        `[GridSpatialIndex] Removing item ${item.id} from:\n` +
-          `  Spatial cells: [${spatialCells.join(", ")}]\n` +
-          `  Projected cells: [${projectedCells.join(", ")}]`,
-      );
-    }
-
-    // Remove from cuboid index
-    this.#removeFromIndex(item, this.#spatialCells, this.#itemToCellKeys);
-    this.#cuboidCellExtentDirty = true;
-
-    // Remove from projected cell index
-    this.#removeFromIndex(
-      item,
-      this.#projectedCells,
-      this.#itemToProjectedCellKeys,
-    );
-
-    // Remove from projection axes map
-    this.#itemToProjectionAxesMap.delete(item);
-
-    if (DEBUG_SPATIAL_INDEX && DEBUG_ITEM_IDS.includes(item.id)) {
-      // Thorough verification that item is completely removed
-      // Check all spatial cells
-      for (const [cellKey, cell] of this.#spatialCells) {
-        if (cell.has(item)) {
-          throw new Error(
-            `Item ${item.id} still found in spatial cell ${cellKey} after removal`,
-          );
-        }
-      }
-
-      // Check all projected cells
-      for (const [cellKey, cell] of this.#projectedCells) {
-        if (cell.has(item)) {
-          throw new Error(
-            `Item ${item.id} still found in projected cell ${cellKey} after removal`,
-          );
-        }
-      }
-
-      // Check reverse mappings
-      if (this.#itemToCellKeys.has(item)) {
-        throw new Error(
-          `Item ${item.id} still in itemToCellKeys map after removal`,
-        );
-      }
-
-      if (this.#itemToProjectedCellKeys.has(item)) {
-        throw new Error(
-          `Item ${item.id} still in itemToProjectedCellKeys map after removal`,
-        );
-      }
-
-      // Check projection axes map
-      if (this.#itemToProjectionAxesMap.has(item)) {
-        throw new Error(
-          `Item ${item.id} still in itemToProjectionAxesMap after removal`,
-        );
-      }
-
-      console.log(
-        `[GridSpatialIndex] Successfully removed item ${item.id} from all indexes`,
-      );
-    }
-  }
-
-  /**
-   * Update an item's position in the spatial index. After this,
-   * collision detection should work
-   */
-  updateItemSpatialIndex(
-    /** The item to update */
-    item: Item,
-  ): void {
-    if (DEBUG_SPATIAL_INDEX && DEBUG_ITEM_IDS.includes(item.id)) {
-      const oldCells = Array.from(this.#itemToCellKeys.get(item) || []);
-      console.log(
-        `[GridSpatialIndex] Updating spatial index for item ${item.id} from cells: [${oldCells.join(", ")}]`,
-      );
-    }
-
-    // Update in cuboid index
-    this.#updateInIndex(
-      item,
-      this.#iterateCuboidCellKeys(item.state.position, item.aabb),
-      this.#spatialCells,
-      this.#itemToCellKeys,
-      (cellKey) => this.#ensureCell(cellKey),
-    );
-    this.#cuboidCellExtentDirty = true;
-
-    if (DEBUG_SPATIAL_INDEX && DEBUG_ITEM_IDS.includes(item.id)) {
-      const newCells = Array.from(this.#itemToCellKeys.get(item) || []);
-      console.log(
-        `[GridSpatialIndex] Updated item ${item.id} to position (${item.state.position.x}, ${item.state.position.y}, ${item.state.position.z}):\n` +
-          `  Now in spatial cells: [${newCells.join(", ")}]`,
-      );
-    }
-  }
-  /**
-   * Update an item's position in the grid. After this, z-edge graph
-   * updates should work.
-   */
-  updateItemProjectedIndex(
-    /** The item to update */
-    item: Item,
-  ): void {
-    if (DEBUG_SPATIAL_INDEX && DEBUG_ITEM_IDS.includes(item.id)) {
-      const oldCells = Array.from(
-        this.#itemToProjectedCellKeys.get(item) || [],
-      );
-      console.log(
-        `[GridSpatialIndex] Updating projected index for item ${item.id} from cells: [${oldCells.join(", ")}]`,
-      );
-    }
-
-    // Update in projected cell index
-    this.#upsertAxesProjections(item);
-    this.#updateInIndex(
-      item,
-      this.#iterateProjectionCellKeys(item),
-      this.#projectedCells,
-      this.#itemToProjectedCellKeys,
-      (cellKey) => this.#ensureProjectedCell(cellKey),
-    );
-
-    if (DEBUG_SPATIAL_INDEX && DEBUG_ITEM_IDS.includes(item.id)) {
-      const newCells = Array.from(
-        this.#itemToProjectedCellKeys.get(item) || [],
-      );
-      console.log(
-        `[GridSpatialIndex] Updated item ${item.id} projected index:\n` +
-          `  Now in projected cells: [${newCells.join(", ")}]`,
-      );
-    }
-  }
-
-  /** world x → cuboid cell x (same flooring as #iterateCuboidCellKeys) */
-  worldXToCellX(worldX: number): number {
-    return Math.floor(worldX / cellWidth);
-  }
-  /** world y → cuboid cell y (same flooring as #iterateCuboidCellKeys) */
-  worldYToCellY(worldY: number): number {
-    return Math.floor(worldY / cellDepth);
-  }
-
-  /**
-   * The live Set of items occupying one cuboid cell, by integer cell coords.
-   * Returns the stored Bin itself (no copy) — the caller must NOT mutate it,
-   * and should not retain it across index mutations. undefined when the cell is
-   * empty (empty cells aren't stored). Cheaper than getCuboidNeighbourhood,
-   * which allocates a fresh Set; intended for hot per-cell walks.
-   */
-  getCellByCellCoords(
-    cellX: number,
-    cellY: number,
-  ): ReadonlySet<Item> | undefined {
-    return this.#spatialCells.get(this.#makeCellKey(cellX, cellY));
-  }
-
-  /**
-   * Min/max occupied cuboid cell in x and y (the index has no z partition, so
-   * no z extent). undefined when the index is empty. Recomputed lazily on read
-   * when dirty by one pass over the cell keys.
-   */
-  getOccupiedCuboidCellExtent():
-    | { minCellX: number; maxCellX: number; minCellY: number; maxCellY: number }
-    | undefined {
-    if (this.#cuboidCellExtentDirty) {
-      let minCellX = Infinity;
-      let maxCellX = -Infinity;
-      let minCellY = Infinity;
-      let maxCellY = -Infinity;
-      for (const key of this.#spatialCells.keys()) {
-        const { x: cx, y: cy } = this.#decodeCellKey(key);
-        minCellX = Math.min(minCellX, cx);
-        maxCellX = Math.max(maxCellX, cx);
-        minCellY = Math.min(minCellY, cy);
-        maxCellY = Math.max(maxCellY, cy);
-      }
-      this.#cuboidCellExtent =
-        maxCellX === -Infinity ? undefined : (
-          { minCellX, maxCellX, minCellY, maxCellY }
-        );
-      this.#cuboidCellExtentDirty = false;
-    }
-    return this.#cuboidCellExtent;
-  }
-
-  /**
-   * Get all items that occupy cells within the given cuboid region.
-   * Each item appears only once in the returned set.
-   */
-  getCuboidNeighbourhood(
-    /** The position of the cuboid to check */
-    position: Xyz,
-    /** The size of the cuboid to check */
-    size: Xyz,
-    /** Optional item to exclude from results (usually the querying item) */
-    excludeItem?: Indexable,
-  ): Set<Item> {
-    const neighbours = new Set<Item>();
-
-    for (const cellKey of this.#iterateCuboidCellKeys(position, size)) {
-      const cell = this.#spatialCells.get(cellKey);
-      if (cell) {
-        for (const neighbour of cell) {
-          if (neighbour !== excludeItem) {
-            neighbours.add(neighbour);
-          }
-        }
-      }
-    }
-
-    return neighbours;
-  }
-
-  /**
-   * Get all items that occupy projected cells overlapping with the given item.
-   * Each item appears only once in the returned set.
-   *
-   * The item *must* be in the index
-   */
-  getProjectedNeighbourhood(
-    /** The item whose projection neighbourhood to check */
-    item: Item,
-    /** Optional item to exclude from results (usually the querying item) */
-    excludeItem?: Item,
-  ): Set<Item> {
-    const neighbours = new Set<Item>();
-
-    // note we don't call this:
-    // this.#upsertAxesProjections(item);
-    // so it is required the item be in the index, and be up to date
-    for (const cellKey of this.#iterateProjectionCellKeys(item)) {
-      const cell = this.#projectedCells.get(cellKey);
-      if (cell) {
-        for (const neighbour of cell) {
-          if (neighbour !== excludeItem) {
-            neighbours.add(neighbour);
-          }
-        }
-      }
-    }
-
-    return neighbours;
-  }
-
-  /**
-   * Get all items that share at least one cell with the given item.
-   * Each item appears only once in the returned set.
-   */
-  getItemCuboidNeighbourhood(
-    /**
-     * NOTE: - there is no requirement for this item to actually be in the index itself
-     */
-    item: Indexable,
-  ): Set<Item> {
-    return this.getCuboidNeighbourhood(item.state.position, item.aabb, item);
-  }
-
-  getItemProjectedNeighbourhood(
-    /**
-     * NOTE: - there is no requirement for this item to actually be in the index itself
-     */
-    item: Item,
-  ): Set<Item> {
-    return this.getProjectedNeighbourhood(item, item);
-  }
-
-  /**
-   * Get the cached axes projections for an item in the index.
-   * Returns undefined if the item is not in the index.
-   */
-  getItemAxesProjections(item: Item): ProjectionOnAxes | undefined {
-    return this.#itemToProjectionAxesMap.get(item);
-  }
-
-  /**
-   * return the item id for every item in every column in a Map-like notation
-   *  eg:
-   *     { (0,0) => [item1, item2, item3], (0,1) => ... }
-   */
-  debugToString(): string {
-    // Sort cell keys for consistent output
-    const sortedCellKeys = Array.from(this.#spatialCells.keys()).sort(
-      (a, b) => a - b,
-    );
-
-    const cellEntries = sortedCellKeys.map((cellKey) => {
-      const cell = this.#spatialCells.get(cellKey)!;
-      const itemIds = Array.from(cell).map((item) => item.id);
-      const { x, y } = this.#decodeCellKey(cellKey);
-      return `  (${x},${y}) [raw ${cellKey}] => [${itemIds.join(", ")}]`;
-    });
-
-    // Sort projected cell keys for consistent output
-    const sortedProjectedCellKeys = Array.from(
-      this.#projectedCells.keys(),
-    ).sort((a, b) => a - b);
-
-    const projectedCellEntries = sortedProjectedCellKeys.map((cellKey) => {
-      const cell = this.#projectedCells.get(cellKey)!;
-      const itemIds = Array.from(cell).map((item) => item.id);
-      const { x, y } = this.#decodeCellKey(cellKey);
-      return `  (${x},${y}) [raw ${cellKey}] => [${itemIds.join(", ")}]`;
-    });
-
-    return `GridSpatialIndex (${this.#spatialCells.size} 2D cells, ${this.#projectedCells.size} projected cells, ${this.#itemToCellKeys.size} items) {
-  2D Cells:
-${cellEntries.join(",\n")}
-  Projected Cells:
-${projectedCellEntries.join(",\n")}
-}`;
-  }
-}
